backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
httpx==0.28.1
anthropic==0.42.0
openai==1.58.1
sendgrid==6.11.0
jinja2==3.1.5
boto3==1.36.2
python-dotenv==1.0.1
tenacity==9.0.0
aiofiles==24.1.0
Pillow==11.1.0
weasyprint==63.1
faker==33.1.0
pdfplumber==0.11.4
python-docx==1.1.2
pytest==8.3.4
pytest-asyncio==0.25.0
